Attempting to delete license from db when deactivated #1648

+++ includes/class-give-license-handler.php
@@ -103,7 +103,7 @@ if ( ! class_exists( 'Give_License' ) ) :
 		 *
 		 * @var    string
 		 */
-		private $api_url      = 'https://givewp.com/edd-sl-api/';
+		private $api_url = 'https://givewp.com/edd-sl-api/';
 
 		/**
 		 * Account URL
@@ -113,7 +113,7 @@ if ( ! class_exists( 'Give_License' ) ) :
 		 *
 		 * @var null|string
 		 */
-		private $account_url  = 'https://givewp.com/my-account/';
+		private $account_url = 'https://givewp.com/my-account/';
 
 		/**
 		 * Ccheckout URL
@@ -133,14 +133,14 @@ if ( ! class_exists( 'Give_License' ) ) :
 		 * @access public
 		 * @since  1.0
 		 *
-		 * @param string  $_file
-		 * @param string  $_item_name
-		 * @param string  $_version
-		 * @param string  $_author
-		 * @param string  $_optname
-		 * @param string  $_api_url
-		 * @param string  $_checkout_url
-		 * @param string  $_account_url
+		 * @param string $_file
+		 * @param string $_item_name
+		 * @param string $_version
+		 * @param string $_author
+		 * @param string $_optname
+		 * @param string $_api_url
+		 * @param string $_checkout_url
+		 * @param string $_account_url
 		 */
 		public function __construct( $_file, $_item_name, $_version, $_author, $_optname = null, $_api_url = null, $_checkout_url = null, $_account_url = null ) {
 
@@ -163,7 +163,6 @@ if ( ! class_exists( 'Give_License' ) ) :
 			$this->auto_updater();
 
 		}
-
 
 
 		/**
@@ -206,7 +205,6 @@ if ( ! class_exists( 'Give_License' ) ) :
 
 			// Updater
 			add_action( 'admin_init', array( $this, 'auto_updater' ), 0 );
-
 			add_action( 'admin_notices', array( $this, 'notices' ) );
 
 			// Check license weekly.
@@ -264,12 +262,12 @@ if ( ! class_exists( 'Give_License' ) ) :
 					'desc'    => '',
 					'type'    => 'license_key',
 					'options' => array(
-						'license'       => get_option( $this->item_shortname . '_license_active' ),
-						'shortname'     => $this->item_shortname,
-						'item_name'     => $this->item_name,
-						'api_url'       => $this->api_url,
-						'checkout_url'  => $this->checkout_url,
-						'account_url'   => $this->account_url
+						'license'      => get_option( $this->item_shortname . '_license_active' ),
+						'shortname'    => $this->item_shortname,
+						'item_name'    => $this->item_name,
+						'api_url'      => $this->api_url,
+						'checkout_url' => $this->checkout_url,
+						'account_url'  => $this->account_url
 					),
 					'size'    => 'regular'
 				)
@@ -333,18 +331,20 @@ if ( ! class_exists( 'Give_License' ) ) :
 			}
 
 			// Allow third party addon developers to handle license activation.
-			if( $this->__is_third_party_addon() ){
+			if ( $this->__is_third_party_addon() ) {
 				do_action( 'give_activate_license', $this );
+
 				return;
 			}
 
 			// Delete previous license setting if a empty license key submitted.
 			if ( empty( $_POST[ $this->item_shortname . '_license_key' ] ) ) {
 				delete_option( $this->item_shortname . '_license_active' );
+
 				return;
 			}
 
-			// Do not simultaneously activate any addon if user want to deactivate any addon.
+			// Do not simultaneously activate add-ons if the user want to deactivate a specific add-on.
 			foreach ( $_POST as $key => $value ) {
 				if ( false !== strpos( $key, 'license_key_deactivate' ) ) {
 					// Don't activate a key when deactivating a different key
@@ -362,7 +362,7 @@ if ( ! class_exists( 'Give_License' ) ) :
 			$license = sanitize_text_field( $_POST[ $this->item_shortname . '_license_key' ] );
 
 			// Bailout.
-			if( empty( $license ) ) {
+			if ( empty( $license ) ) {
 				return;
 			}
 
@@ -430,8 +430,9 @@ if ( ! class_exists( 'Give_License' ) ) :
 			}
 
 			// Allow third party add-on developers to handle license deactivation.
-			if( $this->__is_third_party_addon() ){
+			if ( $this->__is_third_party_addon() ) {
 				do_action( 'give_deactivate_license', $this );
+
 				return;
 			}
 
@@ -464,11 +465,18 @@ if ( ! class_exists( 'Give_License' ) ) :
 				// Decode the license data
 				$license_data = json_decode( wp_remote_retrieve_body( $response ) );
 
-				// Remove license data.
-				delete_option( $this->item_shortname . '_license_active' );
+				// Ensure deactivated successfully.
+				if ( isset( $license_data->success ) ) {
 
-				// Remove license key from subscriptions if exist.
-				$this->__remove_license_key_from_subscriptions();
+					// Remove license data.
+					delete_option( $this->item_shortname . '_license_active' );
+					give_delete_option( $this->item_shortname . '_license_key' );
+
+					// Remove license key from subscriptions if exist.
+					$this->__remove_license_key_from_subscriptions();
+
+				}
+
 			}
 		}
 
@@ -482,27 +490,28 @@ if ( ! class_exists( 'Give_License' ) ) :
 		 */
 		public function weekly_license_check() {
 
-			if( ! empty( $_POST['give_settings'] ) ) {
+			if ( ! empty( $_POST['give_settings'] ) ) {
 				// Don't fire when saving settings
 				return false;
 			}
 
-			if( empty( $this->license ) ) {
+			if ( empty( $this->license ) ) {
 				return false;
 			}
 
 			// Allow third party add-on developers to handle their license check.
-			if( $this->__is_third_party_addon() ){
+			if ( $this->__is_third_party_addon() ) {
 				do_action( 'give_weekly_license_check', $this );
+
 				return false;
 			}
 
 			// Data to send in our API request.
 			$api_params = array(
-				'edd_action'=> 'check_license',
-				'license' 	=> $this->license,
-				'item_name' => urlencode( $this->item_name ),
-				'url'       => home_url()
+				'edd_action' => 'check_license',
+				'license'    => $this->license,
+				'item_name'  => urlencode( $this->item_name ),
+				'url'        => home_url()
 			);
 
 			// Call the API.
@@ -534,24 +543,25 @@ if ( ! class_exists( 'Give_License' ) ) :
 		 */
 		public function weekly_subscription_check() {
 
-			if( ! empty( $_POST['give_settings'] ) ) {
+			if ( ! empty( $_POST['give_settings'] ) ) {
 				// Don't fire when saving settings
 				return false;
 			}
 
 			// Remove old subscription data.
-			if( absint( get_option( '_give_subscriptions_edit_last', true ) ) < current_time( 'timestamp' , 1 ) ){
+			if ( absint( get_option( '_give_subscriptions_edit_last', true ) ) < current_time( 'timestamp', 1 ) ) {
 				delete_option( 'give_subscriptions' );
-				update_option( '_give_subscriptions_edit_last', strtotime( '+ 1 day', current_time( 'timestamp' , 1 ) ) );
+				update_option( '_give_subscriptions_edit_last', strtotime( '+ 1 day', current_time( 'timestamp', 1 ) ) );
 			}
 
-			if( empty( $this->license ) ) {
+			if ( empty( $this->license ) ) {
 				return false;
 			}
 
 			// Allow third party add-on developers to handle their subscription check.
-			if( $this->__is_third_party_addon() ){
+			if ( $this->__is_third_party_addon() ) {
 				do_action( 'give_weekly_subscription_check', $this );
+
 				return false;
 			}
 
@@ -563,10 +573,10 @@ if ( ! class_exists( 'Give_License' ) ) :
 				// Do not get confused with edd_action check_subscription.
 				// By default edd software licensing api does not have api to check subscription.
 				// This is a custom feature to check subscriptions.
-				'edd_action'=> 'check_subscription',
-				'license' 	=> $this->license,
-				'item_name' => urlencode( $this->item_name ),
-				'url'       => home_url()
+				'edd_action' => 'check_subscription',
+				'license'    => $this->license,
+				'item_name'  => urlencode( $this->item_name ),
+				'url'        => home_url()
 			);
 
 			// Call the API
@@ -586,17 +596,17 @@ if ( ! class_exists( 'Give_License' ) ) :
 
 			$subscription_data = json_decode( wp_remote_retrieve_body( $response ), true );
 
-			if( ! empty( $subscription_data['success'] ) && absint( $subscription_data['success'] ) ) {
+			if ( ! empty( $subscription_data['success'] ) && absint( $subscription_data['success'] ) ) {
 				$subscriptions = get_option( 'give_subscriptions', array() );
 
 				// Update subscription data only if subscription does not exist already.
-				if( ! array_key_exists( $subscription_data['id'], $subscriptions ) ) {
-					$subscriptions[ $subscription_data['id'] ] = $subscription_data;
+				if ( ! array_key_exists( $subscription_data['id'], $subscriptions ) ) {
+					$subscriptions[ $subscription_data['id'] ]             = $subscription_data;
 					$subscriptions[ $subscription_data['id'] ]['licenses'] = array();
 				}
 
 				// Store licenses for subscription.
-				if( ! in_array( $this->license, $subscriptions[ $subscription_data['id'] ]['licenses'] ) ) {
+				if ( ! in_array( $this->license, $subscriptions[ $subscription_data['id'] ]['licenses'] ) ) {
 					$subscriptions[ $subscription_data['id'] ]['licenses'][] = $this->license;
 				}
 
@@ -618,7 +628,7 @@ if ( ! class_exists( 'Give_License' ) ) :
 				return false;
 			}
 
-			if( empty( $this->license ) ) {
+			if ( empty( $this->license ) ) {
 				return false;
 			}
 
@@ -627,10 +637,10 @@ if ( ! class_exists( 'Give_License' ) ) :
 				// Do not get confused with edd_action check_subscription.
 				// By default edd software licensing api does not have api to check subscription.
 				// This is a custom feature to check subscriptions.
-				'edd_action'=> 'check_subscription',
-				'license' 	=> $this->license,
-				'item_name' => urlencode( $this->item_name ),
-				'url'       => home_url()
+				'edd_action' => 'check_subscription',
+				'license'    => $this->license,
+				'item_name'  => urlencode( $this->item_name ),
+				'url'        => home_url()
 			);
 
 			// Call the API
@@ -650,17 +660,17 @@ if ( ! class_exists( 'Give_License' ) ) :
 
 			$subscription_data = json_decode( wp_remote_retrieve_body( $response ), true );
 
-			if( ! empty( $subscription_data['success'] ) && absint( $subscription_data['success'] ) ) {
+			if ( ! empty( $subscription_data['success'] ) && absint( $subscription_data['success'] ) ) {
 				$subscriptions = get_option( 'give_subscriptions', array() );
 
 				// Update subscription data only if subscription does not exist already.
-				if( ! array_key_exists( $subscription_data['id'], $subscriptions ) ) {
-					$subscriptions[ $subscription_data['id'] ] = $subscription_data;
+				if ( ! array_key_exists( $subscription_data['id'], $subscriptions ) ) {
+					$subscriptions[ $subscription_data['id'] ]             = $subscription_data;
 					$subscriptions[ $subscription_data['id'] ]['licenses'] = array();
 				}
 
 				// Store licenses for subscription.
-				if( ! in_array( $this->license, $subscriptions[ $subscription_data['id'] ]['licenses'] ) ) {
+				if ( ! in_array( $this->license, $subscriptions[ $subscription_data['id'] ]['licenses'] ) ) {
 					$subscriptions[ $subscription_data['id'] ]['licenses'][] = $this->license;
 				}
 
@@ -678,12 +688,12 @@ if ( ! class_exists( 'Give_License' ) ) :
 		 */
 		public function notices() {
 
-			if( ! current_user_can( 'manage_give_settings' ) ) {
+			if ( ! current_user_can( 'manage_give_settings' ) ) {
 				return;
 			}
 
 			// Do not show licenses notices on license tab.
-			if( 'licenses' === give_get_current_setting_tab() ) {
+			if ( 'licenses' === give_get_current_setting_tab() ) {
 				return;
 			}
 
@@ -693,14 +703,14 @@ if ( ! class_exists( 'Give_License' ) ) :
 
 			// Set default value.
 			$addon_license_key_in_subscriptions = ! empty( $addon_license_key_in_subscriptions ) ? $addon_license_key_in_subscriptions : array();
-			$messages = array();
+			$messages                           = array();
 
-			if(
+			if (
 				empty( $this->license )
 				&& ! $this->__is_notice_dismissed( 'general' )
-			    && empty( $showed_invalid_message )
+				&& empty( $showed_invalid_message )
 			) {
-				$messages['general'] = sprintf(
+				$messages['general']    = sprintf(
 					__( 'You have invalid or expired license keys for one or more Give Add-ons. Please go to the <a href="%s">licenses page</a> to correct this issue.', 'give' ),
 					admin_url( 'edit.php?post_type=give_forms&page=give-settings&tab=licenses' )
 				);
@@ -712,34 +722,34 @@ if ( ! class_exists( 'Give_License' ) ) :
 			$subscriptions = get_option( 'give_subscriptions' );
 
 			// Show subscription messages.
-			if( ! empty( $subscriptions ) && ! $showed_subscriptions_message ) {
+			if ( ! empty( $subscriptions ) && ! $showed_subscriptions_message ) {
 
 				foreach ( $subscriptions as $subscription ) {
 					// Subscription expires timestamp.
 					$subscription_expires = strtotime( $subscription['expires'] );
 
 					// Start showing subscriptions message before one week of renewal date.
-					if( strtotime( '- 7 days', $subscription_expires ) > current_time( 'timestamp', 1 ) ) {
+					if ( strtotime( '- 7 days', $subscription_expires ) > current_time( 'timestamp', 1 ) ) {
 						continue;
 					}
 
 					// Check if subscription message already exist in messages.
-					if( array_key_exists( $subscription['id'], $messages ) ) {
+					if ( array_key_exists( $subscription['id'], $messages ) ) {
 						continue;
 					}
 
-					if( ( ! $this->__is_notice_dismissed( $subscription['id'] ) && 'active' !== $subscription['status'] ) ) {
+					if ( ( ! $this->__is_notice_dismissed( $subscription['id'] ) && 'active' !== $subscription['status'] ) ) {
 
-						if( strtotime( $subscription['expires'] ) < current_time( 'timestamp', 1 ) ) {// Check if license already expired.
-							$messages[$subscription['id']] = sprintf(
+						if ( strtotime( $subscription['expires'] ) < current_time( 'timestamp', 1 ) ) {// Check if license already expired.
+							$messages[ $subscription['id'] ] = sprintf(
 								__( 'Your Give add-on license expired for payment <a href="%s" target="_blank">#%d</a>. <a href="%s" target="_blank">Click to renew an existing license</a> or <a href="%s">Click here if already renewed</a>.', 'give' ),
 								urldecode( $subscription['invoice_url'] ),
 								$subscription['payment_id'],
 								"{$this->checkout_url}?edd_license_key={$subscription['license_key']}&utm_campaign=admin&utm_source=licenses&utm_medium=expired",
 								esc_url( add_query_arg( '_give_hide_license_notices_permanently', $subscription['id'], $_SERVER['REQUEST_URI'] ) )
 							);
-						}else{
-							$messages[$subscription['id']] = sprintf(
+						} else {
+							$messages[ $subscription['id'] ] = sprintf(
 								__( 'Your Give add-on license will expire in %s for payment <a href="%s" target="_blank">#%d</a>. <a href="%s" target="_blank">Click to renew an existing license</a> or <a href="%s">Click here if already renewed</a>.', 'give' ),
 								human_time_diff( current_time( 'timestamp', 1 ), strtotime( $subscription['expires'] ) ),
 								urldecode( $subscription['invoice_url'] ),
@@ -757,14 +767,14 @@ if ( ! class_exists( 'Give_License' ) ) :
 			}
 
 			// Show non subscription addon messages.
-			if(
+			if (
 				! in_array( $this->license, $addon_license_key_in_subscriptions )
 				&& ! $this->__is_notice_dismissed( 'general' )
 				&& ! $this->is_valid_license()
 				&& empty( $showed_invalid_message )
 			) {
 
-				$messages['general'] = sprintf(
+				$messages['general']    = sprintf(
 					__( 'You have invalid or expired license keys for one or more Give Add-ons. Please go to the <a href="%s">licenses page</a> to correct this issue.', 'give' ),
 					admin_url( 'edit.php?post_type=give_forms&page=give-settings&tab=licenses' )
 				);
@@ -773,8 +783,8 @@ if ( ! class_exists( 'Give_License' ) ) :
 			}
 
 			// Print messages.
-			if( ! empty( $messages ) ) {
-				foreach( $messages as $notice_id => $message ) {
+			if ( ! empty( $messages ) ) {
+				foreach ( $messages as $notice_id => $message ) {
 
 					echo sprintf(
 						'<div class="notice notice-error is-dismissible give-license-notice" data-dismiss-notice-shortly="%1$s"><p>%2$s</p></div>',
@@ -794,7 +804,7 @@ if ( ! class_exists( 'Give_License' ) ) :
 		 * @return bool
 		 */
 		public function is_valid_license() {
-			if( apply_filters( 'give_is_valid_license' , ( is_object( $this->license_data ) && ! empty( $this->license_data ) &&  property_exists( $this->license_data, 'license' )&& 'valid' === $this->license_data->license ) ) ) {
+			if ( apply_filters( 'give_is_valid_license', ( is_object( $this->license_data ) && ! empty( $this->license_data ) && property_exists( $this->license_data, 'license' ) && 'valid' === $this->license_data->license ) ) ) {
 				return true;
 			}
 
@@ -824,20 +834,20 @@ if ( ! class_exists( 'Give_License' ) ) :
 		 *
 		 * @return void|bool
 		 */
-		private function __remove_license_key_from_subscriptions(){
+		private function __remove_license_key_from_subscriptions() {
 			$subscriptions = get_option( 'give_subscriptions', array() );
 
 			// Bailout.
-			if( empty( $this->license ) ) {
+			if ( empty( $this->license ) ) {
 				return false;
 			}
 
-			if( ! empty( $subscriptions ) ) {
+			if ( ! empty( $subscriptions ) ) {
 				foreach ( $subscriptions as $subscription_id => $subscription ) {
 					$license_index = array_search( $this->license, $subscription['licenses'] );
-					if( false !== $license_index ) {
+					if ( false !== $license_index ) {
 						// Remove license key.
-						unset( $subscriptions[ $subscription_id ]['licenses'][$license_index] );
+						unset( $subscriptions[ $subscription_id ]['licenses'][ $license_index ] );
 
 						// Rearrange license keys.
 						$subscriptions[ $subscription_id ]['licenses'] = array_values( $subscriptions[ $subscription_id ]['licenses'] );
@@ -858,7 +868,7 @@ if ( ! class_exists( 'Give_License' ) ) :
 		 *
 		 * @return void
 		 */
-		private function __remove_license_notices_show_blocker(){
+		private function __remove_license_notices_show_blocker() {
 			global $wpdb;
 
 			// Delete permanent notice blocker.
@@ -896,8 +906,8 @@ if ( ! class_exists( 'Give_License' ) ) :
 		 *
 		 * @return bool
 		 */
-		private function __is_notice_dismissed( $notice_id ){
-			$current_user = wp_get_current_user();
+		private function __is_notice_dismissed( $notice_id ) {
+			$current_user        = wp_get_current_user();
 			$is_notice_dismissed = false;
 
 			// Ge is notice dismissed permanently.
@@ -906,15 +916,16 @@ if ( ! class_exists( 'Give_License' ) ) :
 				: array();
 
 
-			if(
+			if (
 				in_array( $notice_id, $already_dismiss_notices )
 				|| Give_Cache::get( "_give_hide_license_notices_shortly_{$current_user->ID}_{$notice_id}", true )
 			) {
-				$is_notice_dismissed =  true;
+				$is_notice_dismissed = true;
 			}
 
 			return $is_notice_dismissed;
 		}
+
 	}
 
 endif; // end class_exists check
